requirements.txt
bs4
pyduktape
python-dateutil
requests
